add mockito deps, add BigClassTest example
utilize mockito for dep injection in tests.
I do see how the dep injection allows for testing,
as it separates the initalization from the activity

Activities cannot be tested easily without the android emulator,
meaning in general the more we can test outside of the emulator the
faster, simpler, and more modular our tests become.